now supports dowloading from either git or zenodo

+++ singleuser/start-notebook.d/10_workdir.sh
@@ -1,23 +1,33 @@
 # Copy examples and other "first launch" files over.
 rsync -aq /etc/jupyter/serverroot/ /work/
 
-# Check out community notebooks
-notebooks_dir=/work/notebooks
-if [[ ! -d "$notebooks_dir" ]]; then
-  git clone https://github.com/chameleoncloud/notebooks.git "$notebooks_dir"
-else
-  (cd "$notebooks_dir" && git stash && git pull && git stash pop || true)
-fi
+# Check out community notebooks if not imported
+if [[ "$IS_IMPORTED" = "yes" ]]; then
+    notebooks_dir=/work/notebooks
+    if [[ ! -d "$notebooks_dir" ]]; then
+      git clone https://github.com/chameleoncloud/notebooks.git "$notebooks_dir"
+    else
+      (cd "$notebooks_dir" && git stash && git pull && git stash pop || true)
+    fi
 
-# Our volume mount is at the root directory, link it in to the user's
-# home directory for convenience.
-rm -rf /home/jovyan/work && ln -s /work /home/jovyan/work
+    # Our volume mount is at the root directory, link it in to the user's
+    # home directory for convenience.
+    rm -rf /home/jovyan/work && ln -s /work /home/jovyan/work
+fi
 
 # MAXINE: added the lines below to clone a git repo into the work directory
 experiment_dir=~/work/experiments
 
 if [[ "$IS_IMPORTED" = "yes" ]]; then
-   cd work 
-   git clone $CLONE_URL
-   cd ..
+    cd work 
+    # For Git:
+    if [[ "$IMPORT_SRC" = "git" ]]; then
+        git clone $CLONE_URL
+    # For Zenodo:
+    else
+        wget $ZEN_ZIP 
+        unzip '*.zip'
+        rm *.zip       
+    fi
+    cd ..
 fi
